init autoLoginService
