+++ src/App.js
@@ -1,14 +1,10 @@
 import React, { Suspense, useEffect } from "react";
-import api from "./api/api";
 import activityService from "./services/activityService";
 import conversationService from "./services/conversationsService";
 import messagesService from "./services/messagesService";
-import subscribeForNotifications from "./services/notifications.js";
+import autoLoginService from "./services/autoLoginService.js";
 import { AnimatePresence } from "framer-motion";
 import { Route, Routes, useLocation, useNavigate } from "react-router-dom";
-import { default as EventEmitter } from "./event/eventEmitter";
-import { setSelectedConversation } from "./store/SelectedConversation";
-import { setUserIsLoggedIn } from "./store/UserIsLoggedIn ";
 import { updateNetworkState } from "./store/NetworkState";
 import { useDispatch } from "react-redux";
 
@@ -24,9 +20,9 @@ const Login = React.lazy(() => import("./components/screens/Login"));
 const ErrorPage = React.lazy(() => import("./components/ErrorPage"));
 
 export default function App() {
-  const navigate = useNavigate();
   const location = useLocation();
   const dispatch = useDispatch();
+  const navigate = useNavigate();
 
   useEffect(() => {
     window.addEventListener("offline", () =>
@@ -47,56 +43,14 @@ export default function App() {
     }
 
     const token = localStorage.getItem("sessionId");
-    EventEmitter.subscribe("onConnect", () => {
-      const token = localStorage.getItem("sessionId");
-      if (!token || token === "undefined") {
-        return;
-      }
-      userLoginByToken(token);
-    });
     if (token && token !== "undefined") {
-      userLoginByToken(token);
+      const currentPath = location.hash;
+      navigate(!currentPath ? "/main" : `/main/${currentPath}`);
     } else {
       localStorage.removeItem("sessionId");
       navigate("/login");
     }
   }, []);
-
-  const userLoginByToken = async (token) => {
-    const currentPath = location.hash;
-    dispatch(updateNetworkState(false));
-
-    const handleLoginFailure = () => {
-      localStorage.removeItem("sessionId");
-      navigate("/login");
-      dispatch(updateNetworkState(true));
-      dispatch(setUserIsLoggedIn(false));
-    };
-
-    try {
-      await api.connect();
-      const userToken = await api.userLogin({ token });
-
-      if (userToken && userToken !== "undefined") {
-        localStorage.setItem("sessionId", userToken);
-        subscribeForNotifications();
-
-        if (!currentPath) {
-          navigate("/main");
-        } else {
-          dispatch(setSelectedConversation({ id: currentPath.slice(1) }));
-          navigate(`/main/${currentPath}`);
-        }
-
-        dispatch(updateNetworkState(true));
-        dispatch(setUserIsLoggedIn(true));
-      } else {
-        handleLoginFailure();
-      }
-    } catch (error) {
-      handleLoginFailure();
-    }
-  };
 
   return (
     <Suspense fallback={<PageLoader />}>
