+++ src/components/screens/Login.js
@@ -33,14 +33,12 @@ export default function Login() {
   const onSubmit = async (data) => {
     setLoader(true);
     try {
-      api.connect().then(async () => {
-        const userToken = await api.userLogin(data);
-        localStorage.setItem("sessionId", userToken);
-        navigate("/main");
-        subscribeForNotifications();
-        dispatch(setSelectedConversation({}));
-        dispatch(setUserIsLoggedIn(true));
-      });
+      const userToken = await api.userLogin(data);
+      localStorage.setItem("sessionId", userToken);
+      navigate("/main");
+      subscribeForNotifications();
+      dispatch(setSelectedConversation({}));
+      dispatch(setUserIsLoggedIn(true));
     } catch (error) {
       localStorage.clear();
       alert(error.message);
